requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
python-telegram-bot==21.6
anthropic==0.40.0
boto3==1.35.74
python-multipart==0.0.17
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
httpx==0.28.1
pillow==11.0.0
stripe==11.3.0
aiofiles==24.1.0
